Spork 14 / Max Supply / Seesaw collateral
Added the spork 14 code back into main code base so we can use it to get public on the proper protocol.
Max supply was set to have 5 coins after max supply was reached so changed it to 0 to not ever go over 100m coins.
Seesaw collateral amount was set to 5k coins, changed it to proper 15k collateral.